backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-dotenv==1.0.1
langgraph==0.2.60
langchain-core==0.3.28
langchain-groq==0.2.2
psycopg2-binary==2.9.10
pymysql==1.1.1
